+++ cli/__init__.py
@@ -7,4 +7,4 @@ Lager CLI
 A Command Line Interface for Lager Data
 """
 
-__version__ = '0.11.0'
+__version__ = '0.12.0'